# std.base64: Improve Encoder/Decoder performance