Massively improves performance of encoding and decoding (~5+x faster for unencoded data >= 10 bytes long) via addition of fast loops at start of methods.

Does worsen memory usage of Decoder as now has to store a [4][256]u32 fast lookup array, but likely worthwhile for this performance increase - if wanting to provide a lower memory usage decoder, may be worth having a fast/slow variant.
Idea behind the fast lookup table is simply to store pre-shifted data as an int based on position in 4 byte block, then just need to bitwise OR each byte in block.
This fast lookup table is also currently being instantied for DecoderWithIgnore as this holds a Decoder despite not using the fast lookup - maybe worth separating these two to avoid this. I did attempt to improve ignore decoder using fast lookup table but worsens performance as no way around carrying out byte by byte checks here (unless you want to make some assumptions about ignore character density).

Can do with some thinking about testing still although these additions should be fairly safe as just "encode/decode simple data until need to handle last 4 bytes or find an invalid character".
 
Also wouldn't be surprised if lot more code in this repo could be improved via use of read/writeInt methods for working with multiple bytes, as this is pretty much responsible for entire encoding improvement.

ReleaseFast benchmark results - length is byte length of unencoded input:
```
-------------------------------------------------------
Benchmark                           Time     Iterations
-------------------------------------------------------
LENGTH = 1
ringbuffer/stdBase64Encode      0.473 ns     2920527750
ringbuffer/newEncode            0.471 ns     2968441751
ringbuffer/stdBase64Decode      0.323 ns     4444695492
ringbuffer/newDecode            0.315 ns     4447358763

LENGTH  = 5
ringbuffer/stdBase64Encode      0.577 ns     2482595234
ringbuffer/newEncode            0.575 ns     2433304640
ringbuffer/stdBase64Decode      0.335 ns     4341348888
ringbuffer/newDecode            0.481 ns     2965567120

LENGTH = 10
ringbuffer/stdBase64Encode      0.629 ns     2221372018
ringbuffer/newEncode            0.634 ns     2232671625
ringbuffer/stdBase64Decode       8.17 ns      168490999
ringbuffer/newDecode            0.782 ns     1786256635

LENGTH = 25
ringbuffer/stdBase64Encode       38.0 ns       36765244
ringbuffer/newEncode            0.906 ns     1559694087
ringbuffer/stdBase64Decode       29.9 ns       44287977
ringbuffer/newDecode             7.95 ns      175760696


LENGTH = 100
ringbuffer/stdBase64Encode        157 ns        8190325
ringbuffer/newEncode             29.8 ns       46469789
ringbuffer/stdBase64Decode        132 ns       10556440
ringbuffer/newDecode             32.5 ns       43006574


LENGTH = 500
ringbuffer/stdBase64Encode        770 ns        1391510
ringbuffer/jdzEncode              153 ns        8701279
ringbuffer/stdBase64Decode        678 ns        2066127
ringbuffer/jdzDecode              168 ns        8389837

LENGTH = 1000
ringbuffer/stdBase64Encode       1530 ns         831907
ringbuffer/jdzEncode              315 ns        4509426
ringbuffer/stdBase64Decode       1388 ns        1001302
ringbuffer/jdzDecode              338 ns        4136550
```

will have a look at doing this via vector api soon