@wooster0 2023-10-13
Don't use the LUT if the optimize mode is `ReleaseSmall`? `fast_char_to_index: if (builtin.optimize_mode == .ReleaseSmall) void else [4][256]u32` and then depending on `builtin.optimize_mode == .ReleaseSmall` in other places do it either the old way or with the LUT, if that’s not too much trouble?