@wooster0 2023-10-13
Mhm I kind of assumed it affects binary size too. Don't know how much `ReleaseSmall` is about memory usage.