@joadnacer 2023-10-13
tested locally on my benchmarking repo and seems ReleaseSmall binary size is actually increased by 320 bytes if I add if statements around all fast LUT code to not run in ReleaseSmall mode, which is surprising as would have expected this to result in about 30 lines of code being compiled away
![image (27)](https://github.com/ziglang/zig/assets/51955597/acbacb31-956d-46c4-9785-c2f1d39e17b3)
![image (28)](https://github.com/ziglang/zig/assets/51955597/b095ee14-9d91-41fd-8a52-c8b348a69f2f)

don't think these checks are worthwhile