@wooster0 2023-10-13
Strange, I would try using a top level constant `const use_lut = builtin.optimize_mode != .ReleaseSmall;` maybe in which case the `if`s should be evaluated at compile time when you do `if (use_lut)` instead of run time like it's perhaps doing when you tried it. Otherwise, just forget about it.