@joadnacer 2023-10-14
yes no difference (tried comptime ifs as well), imo speed improvement is more than worth here and good to merge as is