@joadnacer 2023-10-14
this could be swapped to using readIntSlice, and could also use writeIntSlice if https://github.com/ziglang/zig/pull/17526 is approved
however just a readability choice, current impl is fully performant