@andrewrk 2023-10-15
> Mhm I kind of assumed it affects binary size too. Don't know how much `ReleaseSmall` is about memory usage.

It's about binary size, not memory usage.

Nice work, @joadnacer, thank you!